Add photon_prob eff

+++ AnaHistos/myhadd.sh
@@ -1,17 +1,23 @@
 #!/bin/bash
 # Function: Combine root files ten by ten.
 
-#cd "$SPIN/taxi/Run13pp510MinBias/10701/data"
-cd "$SPIN/taxi/Run13pp510ERT/10853/data"
+echo -en "This is for the TAXI files, do you really want to continure? [N/y] "
+read -n 1 yes
+echo
+if [[ "${yes}" != "y" ]] ; then
+  exit 0
+fi
+
+cd "$PLHF/taxi/Run13pp510ERT/12232/data"
 rm -f total.root tmp.root
 
+prename="Pi0PP-"
 files=""
 count=0
 
-#for FILE in DirectPhotonPP-*.root ; do
-#  files="${files} ${FILE}"
+#for FILE in ${prename}*.root ; do
 while read -d " " runnumber ; do
-  files="${files} DirectPhotonPP-${runnumber}.root"
+  files="${files} ${prename}${runnumber}.root"
   (( count++ ))
   if (( "${count}" > "9" )) ; then
     if [[ -f "total.root" ]] ; then
@@ -23,7 +29,7 @@ while read -d " " runnumber ; do
     files=""
     count=0
   fi
-done < "$PLHF/taxi/Run13pp510ERT/runlist.txt"
+done < "/phenix/plhf/zji/taxi/Run13pp510MinBias/runlist.txt"
 
 if [[ -n "${files}" ]] ; then
   if [[ -f "total.root" ]] ; then
@@ -33,3 +39,5 @@ if [[ -n "${files}" ]] ; then
   fi
   mv tmp.root total.root
 fi
+
+mv total.root ${prename}histo.root
